requirements.txt
# Core
langgraph>=0.2.18
pydantic>=2.7.0
typer>=0.12.3
rich>=13.7.1
python-dotenv>=1.0.1
tqdm>=4.66.4

# LLM client (OpenAI-compatible)
openai>=1.40.0

# Tools
numexpr>=2.10.0
tavily-python>=0.3.7

# RAG
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0.post1
numpy>=1.26.4

# Evaluation
pandas>=2.2.2

# Testing
pytest>=8.2.0

# Streamlit
streamlit>=1.37.0

# Django and PostgreSQL
Django>=5.0,<6.0
djangorestframework>=3.15
psycopg[binary]>=3.2
dj-database-url>=2.2